Smooth out screen suspension
