+++ internal/view/epic.go
@@ -34,17 +34,17 @@ func (el EpicList) Render() error {
 		tui.WithSidebarSelectedFunc(navigate(el.Server)),
 		tui.WithContentTableOpts(
 			tui.WithSelectedFunc(navigate(el.Server)),
-			tui.WithViewModeFunc(func(r, c int, d interface{}) (func() interface{}, func(interface{}) error) {
+			tui.WithViewModeFunc(func(r, c int, d interface{}) (func() interface{}, func(interface{}) (string, error)) {
 				dataFn := func() interface{} {
 					issue, _ := api.ProxyGetIssue(api.Client(jira.Config{}), d.(tui.TableData)[r][1])
 					return issue
 				}
-				renderFn := func(i interface{}) error {
+				renderFn := func(i interface{}) (string, error) {
 					out, err := renderer.Render(Issue{Data: i.(*jira.Issue)}.String())
 					if err != nil {
-						return err
+						return "", err
 					}
-					return PagerOut(out)
+					return out, nil
 				}
 				return dataFn, renderFn
 			}),
