+++ internal/view/project.go
@@ -7,6 +7,7 @@ import (
 	"text/tabwriter"
 
 	"github.com/ankitpokhrel/jira-cli/pkg/jira"
+	"github.com/ankitpokhrel/jira-cli/pkg/tui"
 )
 
 // ProjectOption is a functional option to wrap project properties.
@@ -54,7 +55,7 @@ func (p Project) Render() error {
 		}
 	}
 
-	return PagerOut(p.buf.String())
+	return tui.PagerOut(p.buf.String())
 }
 
 func (p Project) header() []string {
