+++ pkg/tui/helper.go
@@ -2,6 +2,10 @@ package tui
 
 import (
 	"bufio"
+	"fmt"
+	"os"
+	"os/exec"
+	"runtime"
 	"strings"
 
 	"github.com/gdamore/tcell/v2"
@@ -49,4 +53,35 @@ func getInfoModal() *primitive.Modal {
 		SetBackgroundColor(tcell.ColorSpecial).
 		SetTextColor(tcell.ColorDefault).
 		SetBorderColor(tcell.ColorDefault)
+}
+
+// GetPager returns configured pager.
+func GetPager() string {
+	if runtime.GOOS == "windows" {
+		return ""
+	}
+	pager := os.Getenv("PAGER")
+	if pager == "" && cmdExists("less") {
+		pager = "less -r"
+	}
+	return pager
+}
+
+// PagerOut outputs to configured pager if possible.
+func PagerOut(out string) error {
+	pager := GetPager()
+	if pager == "" {
+		_, err := fmt.Print(out)
+		return err
+	}
+	pa := strings.Split(pager, " ")
+	cmd := exec.Command(pa[0], pa[1:]...)
+	cmd.Stdin = strings.NewReader(out)
+	cmd.Stdout = os.Stdout
+	return cmd.Run()
+}
+
+func cmdExists(cmd string) bool {
+	_, err := exec.LookPath(cmd)
+	return err == nil
 }
